fix(crypto): avoid macOS text-relocations; fix endian and cast

+++ src/kernel/crt/heap.cpp
@@ -318,7 +318,8 @@ HeapDiagnostics ReXHeap::GetDiagnosticsLocked() const {
     diagnostics.capacity += d.capacity;
     diagnostics.allocated += d.allocated;
     diagnostics.peak_allocated += d.peak_allocated;
-    diagnostics.peak_request_size = std::max(diagnostics.peak_request_size, d.peak_request_size);
+    diagnostics.peak_request_size =
+        std::max(diagnostics.peak_request_size, static_cast<uint64_t>(d.peak_request_size));
     diagnostics.oom_count += d.oom_count;
   }
   return diagnostics;
